docs: use relative links to Markdown files
Summary:
Prior to this patch, the docs used absolute links to the GitHub wiki.
This patch changes links to use relative paths, so that the docs can be
hosted anywhere: for instance, in the GitHub tree/blob view.

Generated with:

    base_url_escaped='https://github\.com/JoshuaWise/better-sqlite3/wiki'
    sed -i -e 's#'"${base_url_escaped}"'/\([A-Za-z0-9-]*\)#./\1.md#g' ./docs/*.md

Test Plan:
Push this to GitHub, then show the `docs/` directory in the tree view.
Click the following representative links to see that they work:

  - a link from the home page when viewing the enclosing directory;
  - a link from one Markdown file to another, like "Benchmark results"
    in "README";
  - a page-internal fragment link, like "new Database()" on "API
    documentation";
  - a cross-page fragment link, like "Statements" on "64-bit integer
    support"; and
  - an external link, like "WAL mode" from "Performance".

Note that `git grep -P '(?<!wikipedia\.org/)wiki(?!pedia\.org)'` returns
no results.

wchargin-branch: docs-relative-links