readme: make links relative
Summary:
This includes links to the documentation (formerly hosted on a GitHub
wiki, now in the repository), and also a link to the LICENSE file.

Test Plan:
Push this branch to GitHub. Open the repository and click each link in
the README; note that it navigates to the correct place.

Note that `git grep -P '(?<!wikipedia\.org/)wiki(?!pedia\.org)'` returns
no results.

Note that `git grep JoshuaWise` returns only links that really should be
absolute (e.g., the Travis status, the `JoshuaWise/integer` repository,
issues on this repository).

wchargin-branch: readme-relative-links