docs: specify version of SQLite used
Summary:
Previously, users were directed to inspect `master:/deps` to infer which
version of SQLite was being used from the name of the tarball file. Now,
this information is stored explicitly in the documentation, which means
that the documentation can be properly self-contained. Because it no
longer needs to link to `master`, it remains internally consistent even
when viewing an old version of the documentation.

Test Plan:
Note that the current tarball is `deps/sqlite-autoconf-3240000.tar.gz`,
corresponding to version 3.24.0 released on 2018-06-04.

wchargin-branch: docs-sqlite-version